Create script command to kick scully off, refactor responses into separate file
